Avoid 'conversion may change value' cases related to Synchronizer
Avoid two 'conversion may change value' warnings in synchronizertask.cpp,
confirming a conversion to `int` because ─as it's stated in the
documentation─ "`qint64 QIODevice::read(char *data, qint64 maxSize)` Reads at
most `maxSize` bytes" and `maxSize` is only 1440 in this case.

Avoid a 'conversion may change value' warning in synchronizer.cpp, using
a `time_t` data type instead of an `int`.

Avoid four 'conversion may change value' warnings in synchronizer.cpp,
confirming conversions to `uint` because a non-Krusader function that is
used (`fromTime_t()`) does not accept data types that are bigger than `uint`.

Revision: https://invent.kde.org/utilities/krusader/-/merge_requests/58

+++ krusader/Synchronizer/synchronizertask.cpp
@@ -172,8 +172,8 @@ void CompareContentTask::localFileCompareCycle()
     int cnt = 0;
 
     while (!leftFile->atEnd() && !rightFile->atEnd()) {
-        int leftBytes = leftFile->read(leftBuffer, sizeof(leftBuffer));
-        int rightBytes = rightFile->read(rightBuffer, sizeof(rightBuffer));
+        int leftBytes = static_cast<int>(leftFile->read(leftBuffer, sizeof(leftBuffer)));
+        int rightBytes = static_cast<int>(rightFile->read(rightBuffer, sizeof(rightBuffer)));
 
         if (leftBytes != rightBytes) {
             different = true;
